requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4
tensorflow==2.15.0
streamlit==1.29.0
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
kagglehub==0.1.4
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.2
requests==2.31.0
jupyter==1.0.0
notebook==7.0.6
